+++ test/test_bugs.ml
@@ -3,7 +3,11 @@
    SPDX-License-Identifier: ISC
   ---------------------------------------------------------------------------*)
 
-let test_mapper_table_bug_14 () =
+open B0_std
+open B0_testing
+
+let test_mapper_table_bug_14 =
+  Test.test "mapper table bug (#14)" @@ fun () ->
   let table =
     "| a | b | c |\n\
      |---|---|---|\n\
@@ -13,12 +17,14 @@ let test_mapper_table_bug_14 () =
   in
   let doc = Cmarkit.Doc.of_string ~layout:true ~strict:false table in
   let mdoc = Cmarkit.Mapper.map_doc (Cmarkit.Mapper.make ()) doc in
-  print_endline "Expectation for mapper table bug #14:\n";
-  print_endline (Cmarkit_commonmark.of_doc mdoc);
+  let mdoc = Cmarkit_commonmark.of_doc mdoc in
+  Snap.lines mdoc @@ __POS_OF__
+    "| a | b | c |\n\
+     |---|---|---|\n\
+     | a | b | c |\n\
+     |   | b | c |\n\
+     |   |   | c |\n";
   ()
 
-let main () =
-  test_mapper_table_bug_14 ();
-  ()
-
-let () = if !Sys.interactive then () else main ()
+let main () = Test.main @@ fun () -> Test.autorun ()
+let () = if !Sys.interactive then () else exit (main ())